+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## NEXT
 - Now all Tile  in the map have you own Paint.
+- Update `tiledjsonreader` to `1.3.3`.
+- Tiled improvements : Now if you set the class of objectlayer to `collision`, all object of this layer will be collision.
 
 ## 3.0.2
 - `BleedingPixel` improvements.